blueprint split of auth pages
