+++ devtools/output.txt
@@ -1,0 +1,112 @@
+/opt/arm/arm/migrations/versions/6dfe7244b18e_.py:9:1: F401 'sqlalchemy as sa' imported but unused
+import sqlalchemy as sa
+^
+/opt/arm/arm/migrations/versions/95623e8c5d58_merge_add_notification_and_create_sys_.py:8:1: F401 'alembic.op' imported but unused
+from alembic import op
+^
+/opt/arm/arm/migrations/versions/95623e8c5d58_merge_add_notification_and_create_sys_.py:9:1: F401 'sqlalchemy as sa' imported but unused
+import sqlalchemy as sa
+^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:22:13: E131 continuation line unaligned for hanging indent
+            sa.Column('id', sa.Integer(), primary_key=True),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:23:13: E131 continuation line unaligned for hanging indent
+            sa.Column('name', sa.String(length=100), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:24:13: E131 continuation line unaligned for hanging indent
+            sa.Column('cpu', sa.String(length=100), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:25:13: E131 continuation line unaligned for hanging indent
+            sa.Column('description', sa.Unicode(length=200), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:26:13: E131 continuation line unaligned for hanging indent
+            sa.Column('mem_total', sa.Float(), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:30:13: E131 continuation line unaligned for hanging indent
+            sa.Column('drive_id', sa.Integer(), primary_key=True),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:31:13: E131 continuation line unaligned for hanging indent
+            sa.Column('name', sa.String(length=100), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:32:13: E131 continuation line unaligned for hanging indent
+            sa.Column('type', sa.String(length=20), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:33:13: E131 continuation line unaligned for hanging indent
+            sa.Column('mount', sa.String(length=100), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:34:13: E131 continuation line unaligned for hanging indent
+            sa.Column('open', sa.Boolean(), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:35:13: E131 continuation line unaligned for hanging indent
+            sa.Column('job_id_current', sa.Integer(), nullable=True),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:36:13: E131 continuation line unaligned for hanging indent
+            sa.Column('job_id_previous', sa.Integer(), nullable=True),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:37:13: E131 continuation line unaligned for hanging indent
+            sa.Column('description', sa.Unicode(length=200), nullable=False),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:38:13: E131 continuation line unaligned for hanging indent
+            sa.ForeignKeyConstraint(['job_id_current'], ['job.job_id'], ),
+            ^
+/opt/arm/arm/migrations/versions/edce886fb30f_create_system_table.py:39:13: E131 continuation line unaligned for hanging indent
+            sa.ForeignKeyConstraint(['job_id_previous'], ['job.job_id'], )
+            ^
+/opt/arm/arm/ui/auth/auth.py:11:1: F401 'os' imported but unused
+import os
+^
+/opt/arm/arm/ui/auth/auth.py:15:42: E261 at least two spaces before inline comment
+from flask_login import LoginManager, login_required, \
+    current_user, login_user, logout_user # noqa: F401
+                                         ^
+/opt/arm/arm/ui/auth/auth.py:19:1: F401 'arm.config.config as cfg' imported but unused
+import arm.config.config as cfg
+^
+/opt/arm/arm/ui/auth/auth.py:23:28: E127 continuation line over-indented for visual indent
+                           template_folder='templates',
+                           ^
+/opt/arm/arm/ui/auth/auth.py:24:28: E127 continuation line over-indented for visual indent
+                           static_folder='../static')
+                           ^
+/opt/arm/arm/ui/auth/auth.py:76:18: F821 undefined name 'DBUpdate'
+        dbform = DBUpdate(request.form)
+                 ^
+/opt/arm/arm/ui/auth/auth.py:77:21: F821 undefined name 'ui_utils'
+        db_update = ui_utils.arm_db_check()
+                    ^
+/opt/arm/arm/ui/logs/logs.py:10:1: F401 'pathlib.PurePath' imported but unused
+from pathlib import Path, PurePath
+^
+/opt/arm/arm/ui/logs/logs.py:15:1: F401 'arm.ui.db' imported but unused
+from arm.ui import app, db
+^
+/opt/arm/arm/ui/logs/logs.py:16:1: F401 'arm.models.models' imported but unused
+from arm.models import models as models
+^
+/opt/arm/arm/ui/logs/logs.py:20:28: E127 continuation line over-indented for visual indent
+                           template_folder='templates',
+                           ^
+/opt/arm/arm/ui/logs/logs.py:21:28: E127 continuation line over-indented for visual indent
+                           static_folder='../static')
+                           ^
+/opt/arm/arm/ui/logs/logs.py:50:15: F821 undefined name 'ValidationError'
+        raise ValidationError
+              ^
+/opt/arm/arm/ui/logs/logs.py:80:16: F821 undefined name 'send_file'
+        return send_file(full_path, as_attachment=True)
+               ^
+/opt/arm/arm/ui/logs/logs.py:83:15: F821 undefined name 'ValidationError'
+        raise ValidationError
+              ^
+/opt/arm/arm/ui/routes.py:7:1: F401 'bcrypt' imported but unused
+import bcrypt
+^
+/opt/arm/arm/ui/routes.py:13:24: E271 multiple spaces after keyword
+from flask_login import  login_required, UserMixin  # noqa: F401
+                       ^
+4     E127 continuation line over-indented for visual indent
+15    E131 continuation line unaligned for hanging indent
+1     E261 at least two spaces before inline comment
+1     E271 multiple spaces after keyword
+9     F401 'sqlalchemy as sa' imported but unused
+5     F821 undefined name 'DBUpdate'
